refactor error-component.jsx structure
