+++ packages/cashier/src/components/error-component/__tests__/error-component.spec.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { screen, render, fireEvent } from '@testing-library/react';
-import ErrorComponent from '../error-component';
+import ErrorComponent from '../error-component.jsx';
 import { Router } from 'react-router-dom';
 import { createBrowserHistory } from 'history';
 import { unmountComponentAtNode } from 'react-dom';
