+++ packages/cashier/src/components/error-component/error-component.jsx
